single-ai/pyproject.toml
[project]
name = "single-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "yikedu",email = "duyike2021@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.12.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.2.1",
    "requests>=2.32.3,<3.0.0",
    "django>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "supabase>=2.13.0,<3.0.0 ; python_version >= \"3.12\" and python_version < \"4.0\""
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

# Enable Pyflakes, pycodestyle, and isort
select = ["E", "F", "I"]
ignore = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["single_ai"]
combine-as-imports = true
